Moved all errors to errors.ts
